# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dispatcher API",
    "version": "1.0.0",
    "description": "Backend for the Dispatcher Expo app — auth, onboarding, drives board, balances."
  },
  "servers": [{ "url": "https://api-production-f4ac.up.railway.app" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/v1/drives": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List drives. status=open|active|history|assigned|mine; completed=true for public completed feed. Items include poster (and assignee when present). Open board: favorited posters float to top with posterIsFavorite.",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Post a drive (title, passenger phone, vehicle class, seats, trip type; optional address + extra info)",
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/v1/drives/{id}": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Drive detail (passenger contact privacy filtered)",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/drives/{id}/applications": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "summary": "List applicants (poster only). Favorited drivers first with isFavorite.",
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Apply to drive",
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/v1/drives/{id}/accept": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Accept applicant (atomic)",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/drives/{id}/unassign": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Unassign and reopen",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/drives/{id}/complete": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Complete drive and create 10% balance",
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/v1/drives/{id}/hide": {
      "post": {
        "security": [{ "bearerAuth": [] }],
        "summary": "Hide completed drive from public feed (poster)",
        "responses": { "200": { "description": "OK" } }
      }
    },
...
  }
}